@@ -1,7 +1,21 @@
 module github.com/reiver/logjam
 
-go 1.20
+go 1.23.4
 
 require github.com/gorilla/websocket v1.5.0
 
-require github.com/gorilla/mux v1.8.0
+require (
+	github.com/gorilla/mux v1.8.0
+	github.com/reiver/go-erorr v0.0.0-20240801233437-8cbde6d1fa3f
+	github.com/reiver/go-path v0.0.0-20240327181650-5f2ee05890d8
+)
+
+require (
+	github.com/reiver/go-etag v0.0.0-20241130123934-1a98d29fda6e
+	github.com/reiver/go-jsonld v0.0.0-20241216235649-e851bd7e3ab6
+)
+
+require (
+	github.com/reiver/go-json v0.0.0-20241213105958-0fd2a5bb4dd8 // indirect
+	github.com/reiver/go-lck v0.0.0-20240808133902-b56df221c39f // indirect
+)
