@@ -1,10 +1,10 @@
 package rooms
 
-import (
-	"github.com/reiver/logjam/models/dto"
-)
-
 type Repository interface {
+	NumRooms() int
+	RoomIDs() []string
+	ForEachRoom(func(*RoomModel)error) error
+
 	DoesRoomExists(id string) bool
 	CreateRoom(id string) error
 	GetRoom(id string) (*RoomModel, error)
@@ -20,13 +20,13 @@ type Repository interface {
 	GetAllMembersId(roomId string, excludeBroadcaster bool) ([]uint64, error)
 	InsertMemberToTree(roomId string, memberId uint64, isGoldGorilla bool) (parentId *uint64, err error)
 	RemoveMember(roomId string, memberId uint64) (wasBroadcaster bool, nodeChildrenIdList []uint64, err error)
-	SetRoomMetaData(roomId string, metaData map[string]interface{}) error
-	GetRoomMetaData(roomId string) (map[string]interface{}, error)
+	SetRoomMetaData(roomId string, metaData map[string]any) error
+	GetRoomMetaData(roomId string) (map[string]any, error)
 	AddMessageToHistory(roomId string, senderId uint64, msg string) error
 	ClearMessageHistory(roomId string) error
 	GetUserByStreamId(roomId string, streamId string) (*MemberModel, error)
 	IsBroadcaster(roomId string, id uint64) (bool, error)
-	GetMembersList(roomId string) ([]dto.MemberDTO, error)
+	GetMembersList(roomId string) ([]MemberDTO, error)
 	GetChildrenIdList(roomId string, id uint64) ([]uint64, error)
 	IsGGInstance(roomId string, id uint64) bool
 	GetRoomGoldGorillaId(roomId string) (*uint64, error)
