@@ -1,17 +1,18 @@
 package main
 
 import (
-	"github.com/reiver/logjam/cfg"
 	"github.com/reiver/logjam/srv/log"
+
+	// This import enables all the HTTP handlers.
 	_ "github.com/reiver/logjam/www"
 )
 
 func main() {
-	logsrv.Info("main", "LogJam ⚡")
+	log := logsrv.Tag("main")
+
+	log.Info("LogJam ⚡")
 	blur()
 
-	app := App{}
-
-	app.Init(cfg.Config)
-	app.Run()
+	log.Info("Here we go…")
+	webserve()
 }
