@@ -1,6 +1,9 @@
 package verboten
 
 import (
+	_ "github.com/reiver/logjam/www/assets"
 	_ "github.com/reiver/logjam/www/goldgorilla"
-	_ "github.com/reiver/logjam/www/ws"
+	_ "github.com/reiver/logjam/www/hapi"
+	_ "github.com/reiver/logjam/www/src"
+	_ "github.com/reiver/logjam/www/FALLBACK"
 )
